+++ __tests__/restful-model-collection-spec.js
@@ -5,7 +5,6 @@ import NylasConnection from '../src/nylas-connection';
 import RestfulModelCollection from '../src/models/restful-model-collection';
 import Draft from '../src/models/draft';
 import Event from '../src/models/event';
-import { Folder, Label } from '../src/models/folder';
 import Thread from '../src/models/thread';
 
 describe('RestfulModelCollection', () => {
@@ -535,65 +534,6 @@ describe('RestfulModelCollection', () => {
           expect(err).toBe(null);
           done();
         });
-      });
-    });
-
-    describe('test folder and label deletion', () => {
-      test('folder should not have a delete request body', done => {
-        testContext.collection = new RestfulModelCollection(
-          Folder,
-          testContext.connection
-        );
-        testContext.item = new Folder(testContext.connection, {
-          id: '123',
-          displayName: 'foo'
-        });
-        testContext.collection.delete(testContext.item, {'display_name': 'not required'});
-        expect(testContext.connection.request).toHaveBeenCalledWith({
-          method: 'DELETE',
-          qs: {},
-          body: {},
-          path: '/folders/123',
-        });
-        done();
-      });
-
-      test('label should have a delete request body', done => {
-        testContext.collection = new RestfulModelCollection(
-          Label,
-          testContext.connection
-        );
-        testContext.item = new Label(testContext.connection, {
-          id: '123',
-          displayName: 'foo'
-        });
-        testContext.collection.delete(testContext.item, {'display_name': 'required'});
-        expect(testContext.connection.request).toHaveBeenCalledWith({
-          method: 'DELETE',
-          qs: {},
-          body: {'display_name': 'required'},
-          path: '/labels/123',
-        });
-        done();
-      });
-
-      test('delete should read displayName from label object if display_name not provided', done => {
-        testContext.collection = new RestfulModelCollection(
-          Label,
-          testContext.connection
-        );
-        testContext.item = new Label(testContext.connection, {
-          id: '123',
-        });
-        testContext.item.displayName = 'foo';
-        testContext.collection.delete(testContext.item);
-        expect(testContext.connection.request).toHaveBeenCalledWith({
-          method: 'DELETE',
-          qs: {},
-          body: {'display_name': 'foo'},
-          path: '/labels/123',
-        });
-        done();
       });
     });