+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update formatting for DeprecationWarnings
 * Update package-lock.json
 * Add support for `/connect/authorize` endpoint
+* Fix draft deletion bug
 
 ### 4.7.0 / 2019-08-20